// frontend/src/app/models/zakazani-pregled.ts
export class ZakazaniPregled{
    pacijent: string;
    lekar: string;
    ime_lekara: string;
    prezime_lekara: string;
    ogranak: string;
    pregled: string;
    pocetak_pregleda: string;
    kraj_pregleda: string;
    formatiran_datum: string;
    formatirano_vreme: string;
    ima_izvestaj: boolean;
}